refactor: extract tab constants from tab-manager-helpers hub
Extract all constants, configuration objects, and the WorkspaceTab data
model from tab-manager-helpers.js into a new tab-constants.js module.

tab-manager-helpers.js now contains only pure helper functions and
re-exports everything from tab-constants.js for backward compatibility.

11 source files that only imported constants now import directly from
tab-constants.js, reducing the hub's fan-in. JSDoc typedef references
updated across 15 files.

Closes #260
